Se sube los cambios de dominio

+++ src/main/java/com/tallerwebi/dominio/RepositorioObjetivo.java
@@ -4,6 +4,6 @@ public interface RepositorioObjetivo {
 
     Objetivo buscarObjetivo(Integer id, String nombre);
     void guardar(Objetivo objetivo);
-    Egreso buscar(Integer metaMonto);
+    Egreso buscar(Double metaMonto);
     void modificar(Objetivo objetivo);
 }
